refactor(ui-state): C1 — unify dialog/notification managers in presentation/
The two managers were near-identical registries (Map<id,handle> + id minting
+ notifyChange + open/close/getSnapshot/routeEvent), differing only in handle
richness. Collapse the shared mechanics into one generic SessionRegistry core
and co-locate everything the presenter privately owns in a presentation/ folder
(idiomatic, like agent-module/).

- New src/modules/presentation/sessions.ts: SessionRegistry base + DialogManager
  (surface/modal tracking + rich handle) and NotificationManager (onEvent only)
  built on it. Exported names (DialogManager/NotificationManager/DialogHandle/
  NotificationHandle) unchanged, so callers only change an import path.
  isModalOpen() now derives from open sessions, dropping the separate modalIds
  bookkeeping.
- Move the presenter + the managers' state-mocks + integration tests into
  presentation/; delete dialog-manager.ts + notification-manager.ts.
- Update ~16 importers to the new paths.

Single source for the registry mechanics; no behavioral change.
