+++ src/modules/creation-module.integration.test.ts
@@ -15,8 +15,8 @@ import { createMockAccessor } from "../boundaries/platform/config.test-utils";
 import { createCreationModule, validateCloneUrl } from "./creation-module";
 import type { CreationModuleDeps } from "./creation-module";
 import type { IntentModule } from "../intents/lib/module";
-import { createMockDialogManager } from "./dialog-manager.state-mock";
-import type { MockDialogHandle } from "./dialog-manager.state-mock";
+import { createMockDialogManager } from "./presentation/dialog-manager.state-mock";
+import type { MockDialogHandle } from "./presentation/dialog-manager.state-mock";
 import type { Dispatcher } from "../intents/lib/dispatcher";
 import type { DialogConfig, DialogSection } from "../shared/dialog-types";
 import type { ConfigAgentType } from "../boundaries/platform/config";
