+++ src/modules/error-report-module.ts
@@ -36,8 +36,8 @@ import { gzipSync } from "node:zlib";
 import type { IntentModule } from "../intents/lib/module";
 import type { DomainEvent } from "../intents/lib/types";
 import type { IDispatcher } from "../intents/lib/dispatcher";
-import type { DialogHandle } from "./dialog-manager";
-import type { UiPresenter } from "./presentation-module";
+import type { DialogHandle } from "./presentation/sessions";
+import type { UiPresenter } from "./presentation/presentation-module";
 import type { FileSystemBoundary } from "../boundaries/platform/filesystem";
 import type { Logging, Logger } from "../boundaries/platform/logging";
 import type { DialogConfig } from "../shared/dialog-types";
